+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `projects update` command to update an existing project
 - `verify $EMAIL_CODE` command to verify users e-mail with the CLI
 - `run` supports --project/-p
+- `init` saves project context instead of app context
 
 ### Fixed
 
@@ -29,6 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - `apps` commands, replaced with `projects` commands
 - `rename` command in favour of `update`
+- Fix `--me` when in a team context
 
 ## [0.4.0] - 2017-08-31